refactor(auth): extract InAppLoginPanel out of PreLaunchAuthPanel
Pure relocation, no behavior change: the in-app login session UI
(URL + Open/Copy, paste-code box, live phase line, Cancel/"Use
terminal instead") moves to its own file so the Armory/Stash surface
(SPEC_INAPP_CLAUDE_OAUTH_LOGIN_2026_08_03.md §3.3 surface 3) can reuse
it instead of a second hand-rolled copy. Decoupled from
PreLaunchAuthPanel's own AuthFlowController/AuthState reducer —
callers now pass `authUrl` directly (the only field of that reducer's
state the component ever read).
