+++ frontend/app/view/agent/components/PreLaunchAuthPanel.tsx
@@ -43,6 +43,7 @@ import {
     type SelectionOutcome,
 } from "../auth";
 import { getCliCatalogEntry } from "../defaults/cli-catalog";
+import { InAppLoginPanel, type InAppLoginPhase } from "./InAppLoginPanel";
 import { registerSeededAccount } from "../flows/register-seeded-account";
 import { runProviderLogin } from "../flows/run-provider-login";
 import type { ProviderDefinition } from "../providers";
@@ -98,12 +99,6 @@ export interface PreLaunchAuthPanelProps {
  *  generic `WaitingPanel` (whose paste row submits a redirect URL to
  *  `auth.submitcallback` — a backend session this flow doesn't have). */
 const PROVIDER_LOGIN_SESSION_ID = "provider-login";
-
-/** Live phase of the in-app login session, surfaced in `InAppLoginPanel`'s
- *  status line. Driven by `startConnect`'s `onTierChange`/`setAuthUrl`
- *  wiring off `runProviderLogin`'s real transitions (spec §3.1's
- *  `spawning → url_captured → waiting_for_authorize → …` states). */
-type InAppLoginPhase = "starting" | "waiting-authorize" | "fallback" | "terminal-polling";
 
 /** UI hooks `startConnect` (a module-level function) uses to drive the
  *  panel-owned in-app-login signals: the live phase line, and the
@@ -297,7 +292,7 @@ export const PreLaunchAuthPanel = (props: PreLaunchAuthPanelProps): JSX.Element 
                     <InAppLoginPanel
                         providerId={props.provider?.id ?? ""}
                         providerLabel={props.provider?.displayName ?? "this provider"}
-                        state={controller.state()}
+                        authUrl={controller.state().authUrl}
                         phase={inAppPhase()}
                         onCancel={() => void controller.cancel()}
                         onUseTerminal={() => {
@@ -860,200 +855,6 @@ const WaitingPanel = (p: {
                 </div>
             </Show>
             <Button onClick={() => p.onCancel()}>Cancel login</Button>
-        </div>
-    );
-};
-
-/** The in-app login session panel (SPEC_INAPP_CLAUDE_OAUTH_LOGIN_2026_08_03.md
- *  §3.1 / §3.3 surface 1) — the launch-modal sibling of AgentDocumentView's
- *  `AuthUrlBox` (same 2-step authorize-then-paste pattern, same
- *  `setProviderAuth` → login-child-stdin code delivery), plus the pieces the
- *  modal flow needs that the composer box doesn't have: a live phase line
- *  (fed from `runProviderLogin`'s real tier transitions), and the explicit
- *  "Use terminal instead" secondary action that replaces the old
- *  auto-launched terminal. The paste step is the FALLBACK — when the user
- *  authorizes in the browser, the CLI detects completion on its own and the
- *  session resolves with nothing pasted (spec §2 item 3). */
-const InAppLoginPanel = (p: {
-    providerId: string;
-    providerLabel: string;
-    state: AuthState;
-    phase: InAppLoginPhase;
-    onCancel: () => void;
-    onUseTerminal: () => void;
-}): JSX.Element => {
-    const [pasteCode, setPasteCode] = createSignal("");
-    const [pasting, setPasting] = createSignal(false);
-    const [pasteResult, setPasteResult] = createSignal<string | null>(null);
-    let inputRef: HTMLInputElement | undefined;
-
-    // Grab focus when the paste input appears (same rationale as
-    // AuthUrlBox: a pasted code must land HERE, not whatever field held
-    // focus before). Deferred a frame so it wins the modal's own focus
-    // management; fires once per URL appearance.
-    createEffect(() => {
-        if (p.state.authUrl) {
-            requestAnimationFrame(() => inputRef?.focus());
-        }
-    });
-
-    const submitCode = async (explicit?: string): Promise<void> => {
-        // Read from the live input element as a fallback (mirrors
-        // AuthUrlBox): if focus desynced and the controlled signal missed
-        // the paste, the DOM value still holds it.
-        const code = (explicit ?? inputRef?.value ?? pasteCode()).trim();
-        if (!code) return;
-        setPasting(true);
-        setPasteResult(null);
-        try {
-            // Delivered to the login child's stdin via the host's
-            // set_provider_auth → CliLoginStdin::write_line plumbing; the
-            // session's completion poll (inside runProviderLogin) then
-            // observes the CLI finishing. SECURITY: the code is single-use
-            // and PKCE-bound to the spawned process (spec §5) — never logged.
-            await getApi().setProviderAuth(p.providerId, code);
-            setPasteResult("Code accepted — signing you in…");
-            setPasteCode("");
-        } catch (err) {
-            const msg = err instanceof Error ? err.message : String(err);
-            setPasteResult(`Error: ${msg}`);
-        } finally {
-            setPasting(false);
-        }
-    };
-
-    const phaseText = (): string => {
-        switch (p.phase) {
-            case "starting":
-                return `Starting ${p.providerLabel} sign-in — requesting a sign-in link…`;
-            case "waiting-authorize":
-                return "Waiting for you to authorize in the browser — we'll detect it automatically.";
-            case "fallback":
-                return "No sign-in link appeared — trying your existing login, then a terminal…";
-            case "terminal-polling":
-                return "A terminal window opened — finish the login there.";
-        }
-    };
-
-    return (
-        <div class="pre-launch-auth-panel-waiting">
-            <div class="pre-launch-auth-panel-waiting-title">
-                🔐 Sign in to {p.providerLabel}
-            </div>
-            <div class="pre-launch-auth-panel-hint">{phaseText()}</div>
-            {/* reagent P1 on PR #2410: gate on phase too, not just authUrl's
-                presence. "Use terminal instead" kills the tier-1 login child
-                (cancelCliLogin) and moves phase to "fallback"/"terminal-
-                polling" WITHOUT clearing the now-dead authUrl — without this
-                gate, the URL/paste box stayed visible for a session that no
-                longer exists. Pasting a code at that point would call
-                setProviderAuth, whose host handler falls through to the
-                non-CLI config-file branch (cli_login_stdin already cleared)
-                and silently writes the code as a plain auth_token while
-                showing "Code accepted…", misleading the user while the real
-                login runs in the separately-opened terminal instead. */}
-            <Show when={p.state.authUrl && (p.phase === "starting" || p.phase === "waiting-authorize")}>
-                <div class="pre-launch-auth-panel-url-label">
-                    1 · Authorize in your browser (it should have opened — if not, use this link):
-                </div>
-                <div class="pre-launch-auth-panel-url-row">
-                    <code
-                        class="pre-launch-auth-panel-url-text"
-                        title={p.state.authUrl}
-                    >
-                        {p.state.authUrl}
-                    </code>
-                    <Button
-                        className="grey solid"
-                        onClick={() => {
-                            try {
-                                getApi().openExternal(p.state.authUrl);
-                            } catch (e) {
-                                console.warn(`[auth-diag] openExternal failed: ${(e as Error)?.message ?? String(e)}`);
-                            }
-                        }}
-                    >
-                        Open
-                    </Button>
-                    <Button
-                        className="grey solid"
-                        onClick={() => {
-                            // CEF clipboard wrapper, not navigator.clipboard —
-                            // SPEC_UNIFIED_CLIPBOARD_2026_05_18.md §3.3.
-                            void clipboardWriteText(p.state.authUrl).catch((err) =>
-                                console.log("clipboard write failed", err),
-                            );
-                        }}
-                    >
-                        Copy
-                    </Button>
-                </div>
-                <div class="pre-launch-auth-panel-url-label">
-                    2 · If the page shows an authorization code, paste it here:
-                </div>
-                <div class="pre-launch-auth-panel-callback-row">
-                    <input
-                        ref={inputRef}
-                        class="pre-launch-auth-panel-url-input"
-                        type="text"
-                        placeholder="Paste the authorization code…"
-                        value={pasteCode()}
-                        onInput={(e) => setPasteCode(e.currentTarget.value)}
-                        onKeyDown={(e) => {
-                            if (e.key === "Enter") void submitCode();
-                        }}
-                        onPaste={(e) => {
-                            // Auto-submit on paste — pasting the code IS the
-                            // intent to submit (same UX as AuthUrlBox).
-                            const text = (e.clipboardData?.getData("text") ?? "").trim();
-                            if (text) {
-                                setPasteCode(text);
-                                void submitCode(text);
-                            }
-                        }}
-                    />
-                    <Button
-                        className="grey solid"
-                        title="Paste from clipboard and submit"
-                        onClick={() => {
-                            clipboardReadText()
-                                .then((text) => {
-                                    const trimmed = (text ?? "").trim();
-                                    if (trimmed) {
-                                        setPasteCode(trimmed);
-                                        void submitCode(trimmed);
-                                    }
-                                })
-                                .catch(() => {
-                                    setPasteResult("Could not read clipboard — paste manually");
-                                });
-                        }}
-                    >
-                        Paste &amp; submit
-                    </Button>
-                    <Button
-                        className="grey solid"
-                        onClick={() => void submitCode()}
-                        disabled={pasting()}
-                    >
-                        {pasting() ? "…" : "Submit"}
-                    </Button>
-                </div>
-                <Show when={pasteResult()}>
-                    <div class="pre-launch-auth-panel-hint">{pasteResult()}</div>
-                </Show>
-            </Show>
-            <div class="pre-launch-auth-panel-inapp-actions">
-                <Button onClick={() => p.onCancel()}>Cancel login</Button>
-                {/* Explicit terminal fallback — never auto-launched (spec
-                    §3.2). Hidden once the flow is already in the terminal
-                    tiers, where the request would only abort a live poll. */}
-                <Show when={p.phase === "starting" || p.phase === "waiting-authorize"}>
-                    <Button className="grey" onClick={() => p.onUseTerminal()}>
-                        Use terminal instead
-                    </Button>
-                </Show>
-            </div>
         </div>
     );
 };
